chore(tools/docs): repo cleanup follow-up — remove stale info.md/.yamllint, relocate manual tools, fix validate_registers CLI
Follow-up cleanup after the repository file-inventory audit (PR #1740).

Deleted:
- info.md (stale; hacs.json render_readme:true makes README.md canonical)
- .yamllint.yaml (unused; pre-commit uses check-yaml, not yamllint)

Moved (git mv, history preserved) into tools/manual/:
- cleanup_old_entities.py, generate_strings.py, sort_registers_json.py
  so top-level tools/ holds only recurring CI/pre-commit validators.
  Fixed parents[1] -> parents[2] repo-root resolution in the two moved
  scripts that derive ROOT from __file__.

Fixed:
- tools/validate_registers.py standalone CLI/pre-commit path: the
  read_planner stub was missing group_registers and plan_group_reads
  that registers/loader.py now imports, so the CLI raised ImportError.
  Both stub attributes added; `python tools/validate_registers.py`
  now exits 0.

Docs/tests:
- Updated tools/README.md, tools/manual/README.md,
  docs/architecture/file_inventory.md, and the file-inventory audit
  (audit no longer reads a bare "0 safe-delete candidates").
- Updated tests/test_cleanup_old_entities.py import path.
- Added a CHANGELOG Internal entry.

No Modbus register addresses/names, entity/unique/service IDs,
translation keys, or config/options-flow behavior changed; no runtime
module touched; no migration files deleted; modbus_helpers.py not
reintroduced.
