+++ docs/architecture/file_inventory.md
@@ -100,7 +100,7 @@ of the Modbus connection and I/O and is assembled from mixins.
 | `config_flow.py` | Thin public HA entrypoint (re-exports `ThesslaGreenConfigFlow`). | Runtime | Never add logic here. | 🟧 | hassfest | Implementation lives in `_config_flow/`. Exists only to satisfy hassfest. |
 | `_config_flow/` | Config & options flow implementation: user/network/schema/validation/device scan/reauth/options. | Runtime | Flow changes only if task requires. | 🟥 | `tests/test_config_flow_*.py` | **Do not change config/options-flow behaviour** unless explicitly required. Scan runs here; result cached into entry options. |
 | `diagnostics.py` | Config-entry diagnostics payload with IP/serial redaction. | Runtime | New diagnostic fields. | 🟧 | `tests/test_diagnostics*.py` | Keep redaction of host/serial/error messages. Diagnostics are status-only. |
-| `strings.json` | Source translation strings (entities, config flow, services, error/status codes). | Runtime | New UI strings. | 🟥 | `check_translations.py`, `generate_strings.py` | **Translation keys are a public contract** — add, do not rename. |
+| `strings.json` | Source translation strings (entities, config flow, services, error/status codes). | Runtime | New UI strings. | 🟥 | `check_translations.py`, `manual/generate_strings.py` | **Translation keys are a public contract** — add, do not rename. |
 | `translations/en.json`, `pl.json` | Localised strings generated from `strings.json`. | Runtime | Regenerate after `strings.json`. | 🟧 | `check_translations.py` | Keep keys in sync across locales. |
 | `brand/` | HA brand icon/logo assets. | Runtime | Branding. | 🟩 | — | — |
 
@@ -109,6 +109,6 @@ of the Modbus connection and I/O and is assembled from mixins.
 | Path | Role | Kind | When to change | Risk | Related tests/tools | Notes / do-not-change |
 |---|---|---|---|---|---|---|
 | `tests/` | Pytest suite (~250 files): entities, coordinator, write/read-back, services, config flow, dependency direction, register/vendor coverage. | Test | With any behaviour change. | 🟧 | needs **Python 3.13** + HA test stack | `test_dependency_direction.py` enforces layer isolation — may be strengthened, never weakened. Do not edit tests just to collect on older Python. |
-| `tools/` | Recurring validators/generators: `check_maintainability.py`, `validate_entity_mappings.py`, `check_translations.py`, `compare_registers_with_reference.py`, `compare_airpack4_vendor_coverage.py`, `validate_registers.py`, `generate_strings.py`, `sort_registers_json.py`, `cleanup_old_entities.py`, `validate_dashboard_entities.py`. | Tool | New checks/generators. | 🟩 | run in CI / pre-commit / manually | Lightweight validators need only `pydantic`/`PyYAML`/`voluptuous`. Generators must not silently change register names/IDs. |
-| `tools/manual/` | One-shot / manual utilities kept out of the automated pipeline: `migrate_register_names.py`, `translate_register_descriptions.py`, `clear_airflow_stats.py`, `delete_stale_branches.sh`. See `tools/manual/README.md`. | Tool | Rarely (re-run a migration). | 🟩 | manual only | Not imported by runtime, tests, CI, or pre-commit. |
+| `tools/` | Recurring validators/generators: `check_maintainability.py`, `validate_entity_mappings.py`, `check_translations.py`, `compare_registers_with_reference.py`, `compare_airpack4_vendor_coverage.py`, `validate_registers.py`, `validate_dashboard_entities.py`. | Tool | New checks/generators. | 🟩 | run in CI / pre-commit / manually | Lightweight validators need only `pydantic`/`PyYAML`/`voluptuous`. Generators must not silently change register names/IDs. |
+| `tools/manual/` | One-shot / manual utilities kept out of the automated pipeline: `migrate_register_names.py`, `translate_register_descriptions.py`, `clear_airflow_stats.py`, `delete_stale_branches.sh`, `sort_registers_json.py`, `generate_strings.py`, `cleanup_old_entities.py`. See `tools/manual/README.md`. | Tool | Rarely (re-run a migration/generator). | 🟩 | manual only (`cleanup_old_entities.py` unit-tested by `tests/test_cleanup_old_entities.py`) | Not imported by runtime, CI, or pre-commit. |
 | `docs/` | Architecture, plans, audits, real-device validation, release docs. | Docs | Documentation updates. | 🟩 | — | `real_device_validation.md` gates the quality scale. Plan docs (`*_plan.md`) are the source of truth for refactors. This inventory + `runtime_flow.md` + `write_path.md` live in `docs/architecture/`. |
