--- info.md
@@ -1,22 +1,0 @@
-# ThesslaGreen Modbus Integration
-
-Pełna integracja Home Assistant dla rekuperatorów ThesslaGreen z komunikacją Modbus TCP/RTU. Pokrywa **wszystkie 200+ rejestry** z oficjalnej dokumentacji.
-
-## Funkcje
-- 📊 **50+ czujników** – temperatury, przepływy, ciśnienia, jakość powietrza, energia i statusy systemu
-- 🔋 **Szacowanie zużycia energii** – wbudowany czujnik mocy i energii całkowitej
-- 🔍 **40+ sensory binarne** – statusy, tryby, wejścia, alarmy i zabezpieczenia
-- 🎛️ **30+ kontrolki** – Climate, Switch, Number, Select i inne do pełnej kontroli urządzenia
-- 🌡️ **Zaawansowana encja Climate** z preset modes i funkcjami specjalnymi
-- 🛠️ **13 serwisów** API do automatyzacji, diagnostyki i konfiguracji
-- 🌿 Systemy **GWC** i **Bypass**, tryby specjalne (OKAP, KOMINEK, WIETRZENIE, PUSTY DOM, BOOST)
-- 🔧 **Diagnostyka i logowanie** – szczegółowe raporty błędów i wydajności
-- 📅 **Harmonogram tygodniowy** – pełna konfiguracja programów czasowych
-- 🔎 **Inteligentne skanowanie** – tylko aktywne encje
-- 🌍 **Wsparcie wielojęzyczne** – polski i angielski
-
-## Instalacja
-1. Dodaj to repozytorium do HACS
-2. Zainstaluj integrację
-3. Uruchom ponownie Home Assistant
-4. Dodaj integrację przez UI
