+++ tools/README.md
@@ -1,18 +1,9 @@
 # Tools
 
-## Sort register JSON file
-
-The register file should remain deterministically ordered by Modbus function code
-and decimal address. Run the helper below to enforce the ordering:
-
-```bash
-python tools/sort_registers_json.py
-```
-
-By default the script operates on
-`custom_components/thessla_green_modbus/registers/thessla_green_registers_full.json`
-but an alternate path may be supplied as an argument. Other helper scripts import
-this module to ensure the sorting logic lives in a single place.
+Recurring validators/generators run by CI, the `validate-registers` pre-commit
+hook, or `CLAUDE.md` §5. One-shot / manual utilities live in
+[`tools/manual/`](manual/README.md) and are **not** part of the automated
+pipeline.
 
 ## Validate register schema (smoke-check)
 
@@ -32,3 +23,20 @@ python tools/check_maintainability.py
 ```
 
 Checks file and function size thresholds (with configurable limits via CLI flags).
+
+## Other recurring validators
+
+| Script | Purpose |
+|---|---|
+| `validate_entity_mappings.py` | Verify entity mappings against the register map. |
+| `check_translations.py` | Check translation keys are present and in sync across locales. |
+| `compare_registers_with_reference.py` | Compare the register map with the vendor reference (`--show-renames`). |
+| `compare_airpack4_vendor_coverage.py` | Regenerate and diff the AirPack4 vendor coverage docs. |
+| `validate_dashboard_entities.py` | Validate entities referenced by `example_dashboard.yaml`. |
+
+## Manual / one-shot tools
+
+The register-JSON sorter (`sort_registers_json.py`), the strings generator
+(`generate_strings.py`), and the entity-registry cleanup helper
+(`cleanup_old_entities.py`) are manual utilities and now live in
+[`tools/manual/`](manual/README.md).
