+++ tools/manual/generate_strings.py
@@ -1,7 +1,7 @@
 import json
 import pathlib
 
-ROOT = pathlib.Path(__file__).resolve().parents[1]
+ROOT = pathlib.Path(__file__).resolve().parents[2]
 REG = (
     ROOT
     / "custom_components"
