+++ tools/manual/sort_registers_json.py
@@ -6,7 +6,7 @@ import argparse
 import json
 from pathlib import Path
 
-ROOT = Path(__file__).resolve().parents[1]
+ROOT = Path(__file__).resolve().parents[2]
 DEFAULT_JSON = (
     ROOT
     / "custom_components"
